+++ tests/unit/IOUtilTest.php
@@ -1,5 +1,7 @@
 <?php
 
+use Goodby\CSV\Export\Standard\Exception\StrictViolationException;
+use paslandau\ArrayUtility\ArrayUtil;
 use paslandau\IOUtility\IOUtil;
 
 class IOUtilTest extends PHPUnit_Framework_TestCase {
@@ -94,7 +96,7 @@ class IOUtilTest extends PHPUnit_Framework_TestCase {
     public function test_copyDirectory(){
         $temp = self::getTmpDirPath();
 
-        $parent = IOUtil::combinePaths($temp,"test/");
+        $parent = IOUtil::combinePaths($temp,"test_copy/");
         IOUtil::createDirectoryIfNotExists($parent);
         $files = [
             "foo.txt" => "foo",
@@ -126,7 +128,7 @@ class IOUtilTest extends PHPUnit_Framework_TestCase {
     public function test_rename(){
         $temp = self::getTmpDirPath();
 
-        $parent = IOUtil::combinePaths($temp,"test/");
+        $parent = IOUtil::combinePaths($temp,"test_rename/");
         IOUtil::createDirectoryIfNotExists($parent);
         $files = [
             "foo.txt" => "foo",
@@ -155,6 +157,109 @@ class IOUtilTest extends PHPUnit_Framework_TestCase {
             $this->assertArrayHasKey($cleanedFile,$files,"File $cleanedFile not expected!");
             $this->assertEquals($files[$cleanedFile],$content,"Content does match between $cleanedFile and $file");
         }
+    }
+
+    public function test_writeCsvFile_readCsvFile()
+    {
+        $temp = self::getTmpDirPath();
+        $parent = IOUtil::combinePaths($temp, "test_writeCsvFile_readCsvFile/");
+        IOUtil::createDirectoryIfNotExists($parent);
+
+        $encoding = mb_internal_encoding();
+        $escape = "\"";
+        $enclosure = "\"";
+        $delimiter = ",";
+        $withHeader = true;
+        $tests = [
+            "empty-array" => [
+                "input" => [],
+            ],
+            "array-one-empty-line" => [
+                "input" => [[]],
+                "expected" => [],
+            ],
+            "array-one-line-numerical-columns" => [
+                "input" => [[0 => "foo", 1 => "bar"]],
+            ],
+            "array-one-line-text-columns" => [
+                "input" => [["foo" => "foo", "bar" => "bar"]],
+            ],
+            "array-one-line-with-null" => [
+                "input" => [["foo" => "foo", "bar" => null]],
+                "expected" => [["foo" => "foo", "bar" => ""]],
+            ],
+            "array-2-lines" => [
+                "input" => [["foo" => "foo", "bar" => "bar"], ["foo" => "baz", "bar" => "test"],]
+            ],
+            "array-2-lines-different-column-names" => [
+                "input" => [["foo" => "foo", "bar" => "bar"], ["foo_diff" => "baz", "bar_diff" => "test"],],
+                "expected" => [["foo" => "foo", "bar" => "bar"], ["foo" => "baz", "bar" => "test"],]
+            ],
+            "array-2-lines-different-column-number-first" => [
+                "input" => [["foo" => "foo", "bar" => "bar", "baz" => "boom"], ["foo" => "baz", "bar" => "test"],],
+                "expected" => StrictViolationException::class
+            ],
+            "array-2-lines-different-column-number-second" => [
+                "input" => [["foo" => "foo", "bar" => "bar"], ["foo" => "baz", "bar" => "test", "baz" => "boom"],],
+                "expected" => StrictViolationException::class
+            ],
+            "array-1-line-escape" => [
+                "input" => [["foo" => "foo", "bar" => "the escape > $escape < "]]
+            ],
+            "array-1-line-enclosure" => [
+                "input" => [["foo" => "foo", "bar" => "the enclosure > $enclosure < "]]
+            ],
+            "array-1-line-delimiter" => [
+                "input" => [["foo" => "foo", "bar" => "the delimiter > $delimiter < "]]
+            ],
+            "array-1-line-crlf" => [
+                "input" => [["foo" => "foo", "bar" => "the crlf > \r\n < "]]
+            ],
+            // https://bugs.php.net/bug.php?id=43225
+            // TODO add this test again when bug is fixed
+//            "array-1-line-slash-double-quote-bug" => [
+//                "input" => [["foo" => "foo", "bar" => "the bug > \\\" < "]]
+//            ],
+        ];
+
+        foreach ($tests as $test => $values) {
+            $input = $values["input"];
+            $expected = $input;
+            if (array_key_exists("expected", $values)) {
+                $expected = $values["expected"];
+            }
+
+            $actual = null;
+            try {
+                $path = IOUtil::combinePaths($parent, $test . ".csv");
+                IOUtil::writeCsvFile($path, $input, $withHeader, $encoding, $delimiter, $enclosure, $escape);
+                $actualCsv = IOUtil::readCsvFile($path, $withHeader, $encoding, $delimiter, $enclosure, $escape);
+                $actual = $actualCsv->toArray(false, true);
+
+            } catch (Exception $e) {
+                $actual = get_class($e);
+            }
+
+            $msg = [
+                "Error at $test:",
+                "WithHeader: " . ($withHeader ? "true" : "false"),
+                "Encoding: " . $encoding,
+                "Delimiter: " . $delimiter,
+                "Enclosure: " . $enclosure,
+                "Escape: " . $escape,
+                "Input: " . json_encode($input),
+                "Excpected: " . json_encode($expected),
+                "Actual: " . json_encode($actual),
+            ];
+            $msg = implode("\n", $msg);
+//                echo $msg . "\n\n\n";
+            if (is_array($expected)) {
+                $this->assertTrue(ArrayUtil::equals($actual, $expected, true, true, true), $msg);
+            } else {
+                $this->assertEquals($expected, $actual, $msg);
+            }
+        }
+        echo "";
     }
 }